--- packages/rspack/tests/t.test.ts
@@ -1,5 +1,0 @@
-import { describeCases } from "./case.template";
-describeCases({
-	name: "t",
-	casePath: "t"
-});
